test(chart tests): upgraded the tests for more specific testing
Created more testing about the performance of the rendering. During the process of testing,
discovered that the vitest doesn't measure the time of synchronous operations. Because of that the
started making test using Date variable of Node. By using it, made clear that the vitest lib was
causing the test to take so long.